+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/video/DefaultVideoSink.java
@@ -132,7 +132,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
   @Override
   public boolean isReady(boolean rendererOtherwiseReady) {
-    return outputSurface == null || videoFrameReleaseControl.isReady(rendererOtherwiseReady);
+    return videoFrameReleaseControl.isReady(rendererOtherwiseReady);
   }
 
   @Override